implemented queroes to get tags with entries
